15_Coursework

+++ 15_Coursework/src/main/kotlin/Product.kt
@@ -2,8 +2,8 @@ open class Product(
     val name: String,
     override var typeProduct: EnumTypeProduct,
     val category: EnumCategory,
-    val timeUnload: Int,
-    val timeLoad: Int,
+    val timeUnload: Long,
+    val timeLoad: Long,
     val weight: Double
 ) : CategoryProduct(category, typeProduct) {
 }
